Refactor: 돈 관련 필드 long타입으로 변경

+++ src/main/java/controller/LottoController.java
@@ -39,7 +39,7 @@ public class LottoController {
 
     private static PurchasePrice getPurchasePrice() {
         OutputView.printPurchaseMessage();
-        int purchasePrice = InputView.getInt();
+        long purchasePrice = InputView.getLong();
         try {
             return PurchasePrice.from(purchasePrice);
         } catch (IllegalArgumentException e) {
